@@ -1,16 +1,18 @@
 {
 	"name": "is-supported-regexp-flag",
 	"version": "1.0.1",
-	"description": "Check whether a RegExp flag is supported. Mostly useful for `y` and `u`.",
+	"description": "Check whether a RegExp flag is supported",
 	"license": "MIT",
 	"repository": "sindresorhus/is-supported-regexp-flag",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=6"
+		"node": ">=12.20"
 	},
 	"scripts": {
 		"test": "xo && ava"
@@ -21,7 +23,6 @@
 	"keywords": [
 		"regexp",
 		"regex",
-		"re",
 		"check",
 		"is",
 		"support",
@@ -30,7 +31,7 @@
 		"flag"
 	],
 	"devDependencies": {
-		"ava": "*",
-		"xo": "*"
+		"ava": "^3.15.0",
+		"xo": "^0.42.0"
 	}
 }
